requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
streamlit>=1.24.0
pillow>=9.5.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.15.0
kaggle>=1.5.13
opencv-python-headless>=4.7.0
wandb>=0.15.0
tqdm>=4.65.0
seaborn>=0.12.0
matplotlib>=3.7.0